refactor: domain 별로 파일 이동
