+++ backend/src/main/java/com/example/backend/domain/problem/Provider.java
@@ -1,6 +1,7 @@
-package com.example.backend.domain;
+package com.example.backend.domain.problem;
 
 
+import com.example.backend.domain.problem.Problem;
 import lombok.Getter;
 import lombok.Setter;
