+++ backend/src/main/java/com/example/backend/repository/ProblemRepository.java
@@ -1,6 +1,6 @@
 package com.example.backend.repository;
 
-import com.example.backend.domain.Problem;
+import com.example.backend.domain.problem.Problem;
 
 import org.springframework.data.jpa.repository.JpaRepository;
