+++ backend/src/main/java/com/example/backend/repository/ProgrammersProblemDetailRepository.java
@@ -1,6 +1,6 @@
 package com.example.backend.repository;
 
-import com.example.backend.domain.ProgrammersProblemDetail;
+import com.example.backend.domain.problem.ProgrammersProblemDetail;
 
 import org.springframework.data.jpa.repository.JpaRepository;
