+++ backend/src/main/java/com/example/backend/repository/SolutionRepository.java
@@ -1,6 +1,6 @@
 package com.example.backend.repository;
 
-import com.example.backend.domain.Solution;
+import com.example.backend.domain.solution.Solution;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SolutionRepository extends JpaRepository<Solution, Long> {
